pyproject.toml
[build-system]
# Build backend for editable/local package installation
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pazule"
version = "2.0.0"
description = "AI mission platform for location and atmosphere verification."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  # Core API runtime (minimal install profile)
  "flask>=2.3.0",
  "flask-cors>=4.0.0",
  "langchain>=0.1.0",
  "langchain-google-genai>=2.0.0",
  "langchain-community>=0.0.10",
  "langchain-openai>=0.0.5",
  "langgraph>=0.0.10",
  "openai>=1.0.0",
  "pillow>=10.0.0",
  "pillow-heif>=0.13.0",
  "python-dotenv>=1.0.0",
  "pyyaml>=6.0",
  "torch>=2.0.0",
  "transformers>=4.30.0",
]

[project.scripts]
docs-audit = "app.legacy.docs_audit.cli:main"

[project.optional-dependencies]
# Model inference and orchestration stack
model = [
  "torch>=2.0.0",
  "transformers>=4.30.0",
  "pillow-heif>=0.13.0",
  "openai>=1.0.0",
  "langchain>=0.1.0",
  "langchain-openai>=0.0.5",
  "langchain-community>=0.0.10",
  "langgraph>=0.0.10",
]
# Local development helpers
dev = [
  "pytest>=8.0.0",
]

[[tool.uv.index]]
# PyTorch CPU-only index to reduce build size (from ~2GB to ~200MB)
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }

[tool.setuptools.packages.find]
# Include the application package tree
where = ["."]
include = ["app*"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "ruff>=0.3.0",
    "pre-commit>=3.5.0",
]
